casting changes made for rune to string

+++ src/gitlab.com/harihara95/mars-rover/rover/direction.go
@@ -19,7 +19,7 @@ var directions = [...] rune {
 
 func (d direction) String() rune { return directions[d] }
 
-func (d direction) int() int { return d }
+func (d direction) int() int { return d.int() }
 
 func getDirConstant(d rune) direction {
 	switch d {
